+++ c_src/modules/evision_backend/abs.h
@@ -15,9 +15,16 @@ static ERL_NIF_TERM evision_cv_mat_abs(ErlNifEnv *env, int argc, const ERL_NIF_T
 
     {
         Mat img;
+        Mat ret;
 
         if (evision_to_safe(env, evision_get_kw(env, erl_terms, "img"), img, ArgInfo("img", 0))) {
-            return evision::nif::ok(env, evision_from(env, Mat(cv::abs(img))));
+            int error_flag = false;
+            ERRWRAP2(ret = Mat(cv::abs(img)), env, error_flag, error_term);
+            if (!error_flag) {
+                return evision::nif::ok(env, evision_from(env, ret));
+            } else {
+                return error_term;
+            }
         }
     }