+++ c_src/modules/evision_backend/add.h
@@ -22,8 +22,13 @@ static ERL_NIF_TERM evision_cv_mat_add(ErlNifEnv *env, int argc, const ERL_NIF_T
         if (evision_to_safe(env, evision_get_kw(env, erl_terms, "l"), l, ArgInfo("l", 0)) &&
             evision_to_safe(env, evision_get_kw(env, erl_terms, "r"), r, ArgInfo("r", 0))) {
             Mat ret;
-            cv::add(l, r, ret, cv::noArray(), -1);
-            return ok(env, evision_from(env, ret));
+            int error_flag = false;
+            ERRWRAP2(cv::add(l, r, ret, cv::noArray(), -1), env, error_flag, error_term);
+            if (!error_flag) {
+                return evision::nif::ok(env, evision_from(env, ret));
+            } else {
+                return error_term;
+            }
         }
     }
 
@@ -52,9 +57,15 @@ static ERL_NIF_TERM evision_cv_mat_add_typed(ErlNifEnv *env, int argc, const ERL
             evision_to_safe(env, evision_get_kw(env, erl_terms, "l"), l, ArgInfo("l", 0))) {
             int type;
             if (!get_binary_type(t, l, 0, type)) return evision::nif::error(env, "not implemented for the given type");
+            
             Mat ret;
-            cv::add(lhs, rhs, ret, cv::noArray(), type);
-            return evision::nif::ok(env, evision_from(env, ret));
+            int error_flag = false;
+            ERRWRAP2(cv::add(lhs, rhs, ret, cv::noArray(), type), env, error_flag, error_term);
+            if (!error_flag) {
+                return evision::nif::ok(env, evision_from(env, ret));
+            } else {
+                return error_term;
+            }
         }
     }