+++ c_src/modules/evision_backend/divide.h
@@ -12,6 +12,7 @@ static ERL_NIF_TERM evision_cv_mat_divide(ErlNifEnv *env, int argc, const ERL_NI
     std::map<std::string, ERL_NIF_TERM> erl_terms;
     int nif_opts_index = 0;
     evision::nif::parse_arg(env, nif_opts_index, argv, erl_terms);
+    int error_flag = false;
 
     {
         Mat l;
@@ -20,12 +21,17 @@ static ERL_NIF_TERM evision_cv_mat_divide(ErlNifEnv *env, int argc, const ERL_NI
         if (evision_to_safe(env, evision_get_kw(env, erl_terms, "l"), l, ArgInfo("l", 0)) &&
             evision_to_safe(env, evision_get_kw(env, erl_terms, "r"), r, ArgInfo("r", 0))) {
             Mat ret;
-            cv::divide(l, r, ret, 1, -1);
-            return evision::nif::ok(env, evision_from(env, ret));
+
+            ERRWRAP2(cv::divide(l, r, ret, 1, -1), env, error_flag, error_term);
+            if (!error_flag) {
+                return evision::nif::ok(env, evision_from(env, ret));
+            } else {
+                return error_term;
+            }
         }
     }
 
-    if (error_term != 0) return error_term;
+    if (error_flag) return error_term;
     else return evision::nif::error(env, "overload resolution failed");
 }