+++ c_src/modules/evision_backend/logical_and.h
@@ -19,8 +19,14 @@ static ERL_NIF_TERM evision_cv_mat_logical_and(ErlNifEnv *env, int argc, const E
 
         if (evision_to_safe(env, evision_get_kw(env, erl_terms, "l"), l, ArgInfo("l", 0)) &&
             evision_to_safe(env, evision_get_kw(env, erl_terms, "r"), r, ArgInfo("r", 0))) {
-            Mat ret = l & r;
-            return evision::nif::ok(env, evision_from(env, ret));
+            Mat ret;
+            int error_flag = false;
+            ERRWRAP2(ret = l & r, env, error_flag, error_term);
+            if (!error_flag) {
+                return evision::nif::ok(env, evision_from(env, ret));
+            } else {
+                return error_term;
+            }
         }
     }